-Implemented support for displaying progress inside of script blocks -Fixed an issue wherein all progress records would not be displayed in scenario Variable -> Object -> Action -Fixed an issue wherein "detecting total number of objects" to stream progress message did not complete properly if PrtgAPI detected streaming progress was not required -Fixed an issue with unit test ProgressGrouper wherein a record could not be associated its parent if its parent record completed due to being interrupted by the progress of cmdlets inside a script block -Implemented debug log for unit test progress grouper -Fixed an issue when deserializing channel decimal places, wherein PRTG allows decimal values to be entered in this field

+++ PrtgAPI/Objects/Channel.cs
@@ -178,7 +178,7 @@ namespace PrtgAPI
         [Undocumented]
         [XmlElement("injected_decimaldigits")]
         [PropertyParameter(nameof(ChannelProperty.DecimalPlaces))]
-        public int? DecimalPlaces { get; set; } //Automatic, All or a value you specify - todo
+        public double? DecimalPlaces { get; set; } //Automatic, All or a value you specify - todo
 
         /// <summary>
         /// Whether values that are too high or too low should be filtered out of graphs and tables.
